add email method

+++ exterior_design/exterior_design.controller.js
@@ -1,14 +1,20 @@
 const { ExteriorDesignDAO } = require('./exterior_design.DAO')
 const sharp = require('sharp')
 const fs = require('fs')
+const dotenv = require('dotenv')
+dotenv.config()
+
 
 class ExteriorDesignController {
     async createExteriorDesign(req, res) {
-        const { title, desc } = req.body
-        // await sharp(req.file.path)
+        const { title, desc, jwt} = req.body
+        if (!jwt || jwt !== process.env.JWT_TOKEN) {
+            res.status(403).send({ message: 'Ivalid JWT'})
+        } else {
+             // await sharp(req.file.path)
         //     .toFile(`./static/exteriors/${req.file.originalname}`)
         await sharp(req.file.path)
-            .toFile(`/usr/src/app/static/exteriors/${req.file.originalname}`)
+        .toFile(`/usr/src/app/static/exteriors/${req.file.originalname}`)
 
         // const url = `http://localhost:8000/static/exteriors/${req.file.originalname}`
         const url = `https://frolfasd.ru/static/exteriors/${req.file.originalname}`
@@ -26,6 +32,7 @@ class ExteriorDesignController {
                     res.status(400).send({ status: 'Bad Request', message: error.message })
                 }
             });
+        }
     }
 
     async getExteriorDesign(req, res) {
@@ -61,7 +68,11 @@ class ExteriorDesignController {
 
     async deleteExteriorDesign(req, res) {
         const id = req.params.id //id - из url страницы
-        ExteriorDesignDAO.deleteById(id)
+        const { jwt } = req.body
+        if (!jwt || jwt !== process.env.JWT_TOKEN) {
+            res.status(403).send({ message: 'Ivalid JWT'})
+        } else {
+            ExteriorDesignDAO.deleteById(id)
             .then((data) => {
                 res.json('Запись удалена из БД !')
             })
@@ -73,59 +84,66 @@ class ExteriorDesignController {
                 } else {
                     res.status(400).send({ status: 'Bad Request', message: error.message })
                 }
-            });
+            }); 
+        }
+       
     }
 
     async updateExteriorDesign(req, res) {
-        const  { title, desc, imgUrl, isFileChanged } = req.body
+        const  { title, desc, imgUrl, isFileChanged, jwt } = req.body
         const { id } = req.params
-        let deletingFilePath = ''
-        if (imgUrl) {
-            const searchString = "ru/";
-            const startIndex = imgUrl.indexOf(searchString) + searchString.length;
-            deletingFilePath = imgUrl.substring(startIndex);
-        }
-        if (isFileChanged == 1) {
-            fs.unlink(deletingFilePath, () => { // Для удаления cтарых файлов
-                console.log(deletingFilePath)
-            })
-            // await sharp(req.file.path)
-            //     .toFile(`./static/exteriors/${req.file.originalname}`)
-            await sharp(req.file.path)
-                .toFile(`/usr/src/app/static/exteriors/${req.file.originalname}`)
-            const url = `https://frolfasd.ru/static/exteriors/${req.file.originalname}`
-            // const url = `http://localhost:8000/static/exteriors/${req.file.originalname}`
-            fs.unlink(req.file.path, () => { // Для удаления закодированных файлов после использования
-                console.log(req.file.path)
-            })
-            ExteriorDesignDAO.updateById(id, title, url, desc)
-                .then((data) => {
-                    res.json(data)
-                })
-                .catch((error) => {
-                    if (error.status === 404) {
-                        res.status(error.status).send({ status: 'Not found', message: error.message })
-                    } else if (error.status === 500) {
-                        res.status(500).send({ status: 'Problem', message: 'Problem with database' })
-                    } else {
-                        res.status(400).send({ status: 'Bad Request', message: error.message })
-                    }
-                });
+        if (!jwt || jwt !== process.env.JWT_TOKEN) {
+            res.status(403).send({ message: 'Ivalid JWT'})
         } else {
-            ExteriorDesignDAO.updateByIdWithoutUrl(id, title, desc)
-                .then((data) => {
-                    res.json(data)
+            let deletingFilePath = ''
+            if (imgUrl) {
+                const searchString = "ru/";
+                const startIndex = imgUrl.indexOf(searchString) + searchString.length;
+                deletingFilePath = imgUrl.substring(startIndex);
+            }
+            if (isFileChanged == 1) {
+                fs.unlink(deletingFilePath, () => { // Для удаления cтарых файлов
+                    console.log(deletingFilePath)
                 })
-                .catch((error) => {
-                    if (error.status === 404) {
-                        res.status(error.status).send({ status: 'Not found', message: error.message })
-                    } else if (error.status === 500) {
-                        res.status(500).send({ status: 'Problem', message: 'Problem with database' })
-                    } else {
-                        res.status(400).send({ status: 'Bad Request', message: error.message })
-                    }
-                });
+                // await sharp(req.file.path)
+                //     .toFile(`./static/exteriors/${req.file.originalname}`)
+                await sharp(req.file.path)
+                    .toFile(`/usr/src/app/static/exteriors/${req.file.originalname}`)
+                const url = `https://frolfasd.ru/static/exteriors/${req.file.originalname}`
+                // const url = `http://localhost:8000/static/exteriors/${req.file.originalname}`
+                fs.unlink(req.file.path, () => { // Для удаления закодированных файлов после использования
+                    console.log(req.file.path)
+                })
+                ExteriorDesignDAO.updateById(id, title, url, desc)
+                    .then((data) => {
+                        res.json(data)
+                    })
+                    .catch((error) => {
+                        if (error.status === 404) {
+                            res.status(error.status).send({ status: 'Not found', message: error.message })
+                        } else if (error.status === 500) {
+                            res.status(500).send({ status: 'Problem', message: 'Problem with database' })
+                        } else {
+                            res.status(400).send({ status: 'Bad Request', message: error.message })
+                        }
+                    });
+            } else {
+                ExteriorDesignDAO.updateByIdWithoutUrl(id, title, desc)
+                    .then((data) => {
+                        res.json(data)
+                    })
+                    .catch((error) => {
+                        if (error.status === 404) {
+                            res.status(error.status).send({ status: 'Not found', message: error.message })
+                        } else if (error.status === 500) {
+                            res.status(500).send({ status: 'Problem', message: 'Problem with database' })
+                        } else {
+                            res.status(400).send({ status: 'Bad Request', message: error.message })
+                        }
+                    });
+            }
         }
+        
     }
 
     async getImportant(req, res) {
@@ -143,9 +161,12 @@ class ExteriorDesignController {
     }
 
     async updateImportant(req, res) {
-        const { isImportant } = req.body
+        const { isImportant, jwt } = req.body
         const { id } = req.params;
-        ExteriorDesignDAO.updateImportant(id, isImportant)
+        if (!jwt || jwt !== process.env.JWT_TOKEN) {
+            res.status(403).send({ message: 'Ivalid JWT'})
+        } else {
+            ExteriorDesignDAO.updateImportant(id, isImportant)
             .then((data) => {
                 res.json(data)
             })
@@ -158,6 +179,7 @@ class ExteriorDesignController {
                     res.status(400).send({ status: 'Bad Request', message: error.message })
                 }
             });
+        }
     }
 }
